Simplified commonly used code with variables
For CommandBehavior, I put the command user and channel that the command was send inside variables for easy useagle, and I made the Messagebuilder "bob" a global variable named "build".
I then used the MessageBuilder's appendCode command instead of typing out the "```"'s for the code block in the copy command code block.
Finally, I made a listSpeak method inside BotSpeak that'll help with the next code restructure.