Updated Multiple Classes and Variables
For CommandBehavior, I added comCheck which holds the command part of the event's message (up to ~19 characters). The comCheck variable can also be sent to other classes easily if they ever need the exact type of the command sent., like the ReactionList class in the future.
For ReactionList, I changed Role auctionRole to an updated, custom class RoleObject, and changed the variable name to giveawayRole. I also changed much of the private variables to protected.
For the RoleObject class, I changed the variables and made two constructors so this class can be a dual-purpose role information holder: one constructor for reaction roles and another constructor's for plain command roles.