Added method for changing basic roles
I made plainRoleCommand() to divide up the task of giving/removing roles and sending response messages, and I made a second RoleObject to test it out. the commandCheck and the RoleObject's command are checked to see if they're a match, and this comparison will probably be done with an arraylist of RoleObjects later on.